requirements.txt
numpy==1.26.4
pandas==2.2.3
torch==2.4.0
torchvision==0.19.0
opencv-python==4.11.0.86
Pillow==11.2.1
mediapipe==0.10.21
tqdm==4.67.1
pyyaml==6.0.2